refactor(fleet-console): consolidate client modules by boundary
Merge the operation close and undo helpers, the focus guards, the canvas
overlays and geometry, the codex utils, the i18n UI helpers and the sidebar
interaction modules into the files that own them.

The operation lifecycle group is left unmerged: two absorbed modules each
declare a module-scoped listeners set for a different subscriber registry, and
reconciling them would require renaming one.
