+++ runtime/fleet-console/core/client/src/codex/components/toc-sheet.ts
@@ -3,7 +3,7 @@ import type { TocItem } from "@fleet-console/markdown/core";
 import { getGlobalSettingsStoreState } from "../../global-settings-store.js";
 import { getT } from "../../i18n/index.js";
 import { resolveConsoleLanguage } from "../../whatsnew-i18n.js";
-import { escapeAttribute, escapeHtml } from "../utils/html.js";
+import { escapeAttribute, escapeHtml } from "../utils.js";
 
 function resolveActiveLocale() {
   const preference = getGlobalSettingsStoreState().state?.language ?? "auto";
