--- runtime/fleet-console/core/client/src/components/command-band-focus.ts
@@ -1,5 +1,0 @@
-export function focusCommandBandToggleWhenPanelContainsActiveElement(panel: HTMLElement | null, toggleSelector: string): void {
-  const activeElement = document.activeElement;
-  if (panel === null || !(activeElement instanceof Node) || !panel.contains(activeElement)) return;
-  document.querySelector<HTMLButtonElement>(toggleSelector)?.focus();
-}
