+++ runtime/fleet-console/core/client/src/i18n/index.ts
@@ -28,7 +28,6 @@ export function getT(locale: ConsoleLocale): Translate<CoreMessageKey> {
 
 export type { CoreMessageKey };
 export { CORE_MESSAGES } from "./messages/index.js";
-export { diagramHydratorLabels, markdownCopyOptions } from "./markdown-ui.js";
+export { diagramHydratorLabels, markdownCopyOptions } from "./ui.js";
 export { formatAbsoluteDateTime, formatDate, formatRelativeTime } from "./format.js";
-export { renderMessage } from "./rich.js";
-export { translateServerError } from "./server-errors.js";
+export { renderMessage, translateServerError } from "./ui.js";
