+++ runtime/fleet-console/core/client/src/sidebar/operations-side-bar-chip.tsx
@@ -9,7 +9,7 @@ import type { OperationNode } from "../types.js";
 import {
   subscribeSideBarOperationAction,
   type SideBarOperationMenuAction,
-} from "./operation-action-request.js";
+} from "./interaction.js";
 
 export interface SideBarEntry {
   readonly operation: OperationNode;
