+++ runtime/fleet-console/tests/command-band-focus.test.ts
@@ -2,7 +2,7 @@
 
 import { afterEach, beforeEach, describe, expect, it } from "vitest";
 
-import { focusCommandBandToggleWhenPanelContainsActiveElement } from "../core/client/src/components/command-band-focus.js";
+import { focusCommandBandToggleWhenPanelContainsActiveElement } from "../core/client/src/focus-guards.js";
 
 beforeEach(() => document.body.replaceChildren());
 afterEach(() => document.body.replaceChildren());
