+++ runtime/fleet-console/tests/operations-side-bar-status-axis.test.tsx
@@ -7,7 +7,7 @@ import { afterEach, beforeEach, describe, expect, it, vi } from "vitest";
 
 import { getSnapshot, loadForTheater, setOperationOrder } from "../core/client/src/canvas/canvas-store.js";
 import { getIdleArrivalIds, markIdleArrival } from "../core/client/src/operation-idle-arrival.js";
-import { requestSideBarOperationAction } from "../core/client/src/sidebar/operation-action-request.js";
+import { requestSideBarOperationAction } from "../core/client/src/sidebar/interaction.js";
 import { OperationsSideBar } from "../core/client/src/sidebar/operations-side-bar.js";
 import {
   getSideBarStatusSectionCollapsed,
